+++ core/test/Config.test.ts
@@ -102,23 +102,9 @@ describe('get()', () => {
     expect(configContent.stages).toBeUndefined();
   });
 
-  test('should throw an error if a plugin is not an instance of JovoCliPlugin', () => {
-    jest.spyOn(Config.prototype, 'getContent').mockReturnValue({
-      endpoint: 'test',
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore
-      plugins: ['invalid'],
-      stages: { dev: { endpoint: 'dev' } },
-    });
-    jest.spyOn(Config.prototype, 'get').mockReturnValueOnce({});
-
-    const config: Config = new Config('', 'dev');
-    expect(config.get.bind(config)).toThrow('Plugin invalid is not an instance of JovoCliPlugin.');
-  });
-
   test('should merge and return the config with merged plugins for the provided stage', () => {
     const stagedPlugin: Plugin = new Plugin({ files: { foo2: 'bar2' } });
-    stagedPlugin.$id = 'stagedCliPlugin';
+    stagedPlugin.id = 'stagedCliPlugin';
 
     jest.spyOn(Config.prototype, 'getContent').mockReturnValue({
       plugins: [new Plugin({ files: { foo1: 'bar1' } })],
@@ -137,11 +123,11 @@ describe('get()', () => {
     expect(configContent.stages).toBeUndefined();
     expect(configContent).toHaveProperty('plugins');
     expect(configContent.plugins).toHaveLength(1);
-    expect(configContent.plugins![0].$id).toMatch('stagedCliPlugin');
-    expect(configContent.plugins![0].$config.files).toHaveProperty('foo1');
-    expect(configContent.plugins![0].$config.files!.foo1).toMatch('bar1');
-    expect(configContent.plugins![0].$config.files).toHaveProperty('foo2');
-    expect(configContent.plugins![0].$config.files!.foo2).toMatch('bar2');
+    expect(configContent.plugins![0].id).toMatch('stagedCliPlugin');
+    expect(configContent.plugins![0].config.files).toHaveProperty('foo1');
+    expect(configContent.plugins![0].config.files!.foo1).toMatch('bar1');
+    expect(configContent.plugins![0].config.files).toHaveProperty('foo2');
+    expect(configContent.plugins![0].config.files!.foo2).toMatch('bar2');
   });
 });