+++ commands/command-deploy/CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
-## [Unreleased](https://github.com/jovotech/jovo-cli/compare/@jovotech/cli-command-deploy@4.0.0-beta.2...HEAD)
+## [Unreleased](https://github.com/jovotech/jovo-cli/compare/@jovotech/cli-command-deploy@4.0.0-beta.3...HEAD)
+
+> 13 September 2021
+
+### Merged
+- 🚑️ Fix installation of CLI commands [#237](https://github.com/jovotech/jovo-cli/pull/237)
+- ✨ Improve Model Validation [#235](https://github.com/jovotech/jovo-cli/pull/235)
+
+### Commits 
+- :recycle: Refactor model validation [#37d1a13](https://github.com/jovotech/jovo-cli/commit/37d1a13008e3d77c9b92a6d5e9807523f9851127)
+- :heavy_plus_sign: Add CLI command dependencies [#ab94de5](https://github.com/jovotech/jovo-cli/commit/ab94de5cb63d36ff6daec644fc60b6a9aaef3d4c)
+- :fire: Delete package-lock.json [#2a1b1e8](https://github.com/jovotech/jovo-cli/commit/2a1b1e8bc9c6f93c3ce99509451ddda10da5e0b2)
+
+## [@jovotech/cli-command-deploy@4.0.0-beta.3](https://github.com/jovotech/jovo-cli/compare/@jovotech/cli-command-deploy@4.0.0-beta.2...@jovotech/cli-command-deploy@4.0.0-beta.3)
 
 > 27 August 2021